% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shrink_TVP_res.R
\name{print.shrinkTVP}
\alias{print.shrinkTVP}
\title{Nicer printing of shrinkTVP objects}
\usage{
\method{print}{shrinkTVP}(x, ...)
}
\arguments{
\item{x}{A \code{shrinkTVP} object}

\item{...}{Currently ignored.}
}
\value{
Called for its side effects and returns invisibly.
}
\description{
Nicer printing of shrinkTVP objects
}
