% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shrink_TVP_res.R
\name{print.shrinkTVP_res}
\alias{print.shrinkTVP_res}
\title{Nicer printing of shrinkTVP_res objects}
\usage{
\method{print}{shrinkTVP_res}(x, ...)
}
\arguments{
\item{x}{A \code{shrinkTVP_res} object}

\item{...}{Currently ignored.}
}
\value{
Called for its side effects and returns invisibly.
}
\description{
Nicer printing of shrinkTVP_res objects
}
