% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkTVP_methods.R
\name{print.shrinkTVP}
\alias{print.shrinkTVP}
\title{Nicer printing of shrinkTVP objects}
\usage{
\method{print}{shrinkTVP}(x, ...)
}
\arguments{
\item{x}{a \code{shrinkTVP} object.}

\item{...}{Currently ignored.}
}
\value{
Called for its side effects and returns invisibly.
}
\description{
Nicer printing of shrinkTVP objects
}
\author{
Peter Knaus \email{peter.knaus@wu.ac.at}
}
